+++ src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter } from "react-router-dom";
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
-import { About, Hero, Experience, Feedbacks, Navbar, Works, StarsCanvas ,TabLinks,BlogPage
+import { TabLinks, Hero, Navbar, StarsCanvas ,BlogPage
 } from "./components";
 
 import Home from './pages/Home';
@@ -19,14 +19,13 @@ const App = () => {
       <Routes>
         <Route path="/Roger28-51-emr/" element={<Home />} />
 
-        <Route path="/Roger28-51-emr/projects" element={<Experience />} />
+        {/* <Route path="/Roger28-51-emr/blog" element={<BlogPage />} /> */}
 
         <Route path="/Roger28-51-emr/gallery" element={<GalleryComp/>} />
       </Routes>
     </BrowserRouter>
 
 
-   
 
   );
 }
