+++ src/components/Loader.jsx
@@ -1,3 +1,6 @@
+
+///**************** ORIGINAL ************************* */
+
 // import { Html, useProgress } from "@react-three/drei";
 
 // const CanvasLoader = () => {
@@ -39,82 +42,6 @@
 
 
 
-
-// import React, { useState, useEffect } from "react";
-// import { Html } from "@react-three/drei";
-// import loaderImage from "../assets/images/emr_logo_white.png";
-
-
-// import "../assets/css/loader_image.css"
-
-
-// const CanvasLoader = () => {
-//   // State to track if the loader should be shown
-//   const [showLoader, setShowLoader] = useState(true);
-
-//   // Function to hide the loader after 10 seconds
-//   useEffect(() => {
-//     const timer = setTimeout(() => {
-//       setShowLoader(false);
-//     }, 10000);
-//     return () => clearTimeout(timer); // Clear the timer when the component unmounts
-//   }, []);
-
-//   return (
-//     <>
-//       {showLoader && (
-//         <Html
-//           as="div"
-//           center
-//           style={{
-//             display: "flex",
-//             justifyContent: "center",
-//             alignItems: "center",
-//             flexDirection: "column",
-//           }}
-//         >
-//           <span className="canvas-loader"></span>
-//           <p
-//             style={{
-//               fontSize: 20,
-//               color: "#F1F1F1",
-//               fontWeight: 800,
-//               marginTop: 40,
-//             }}
-//           > <img
-//           className="altImage"
-//           src={loaderImage}
-//           alt="Loading..."
-//           style={{
-//               // width: "700px", // Customize the image size as needed
-//               // height: "400px",
-//               // Customize the image size as needed
-//           }}
-//          />
-//             ......................................................
-//           </p>
-//         </Html>
-//       )}
-//     </>
-//   );
-// };
-
-// export default CanvasLoader;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import React, { useState } from "react";
 import { Html } from "@react-three/drei";
 import loaderImage from "../assets/images/emr_logo_white.png";
@@ -122,18 +49,9 @@ import loaderImage from "../assets/images/emr_logo_white.png";
 import "../assets/css/loader_image.css";
 
 const CanvasLoader = () => {
-  // State to track if the content has loaded
+
   const [contentLoaded, setContentLoaded] = useState(false);
-
-  // Simulate content loading (replace this with actual loading logic)
-  // For example, fetch data, load resources, etc.
-  // Once the content is loaded, call setContentLoaded(true)
-
-  // For demonstration purposes, let's simulate content loading with a button click
   const simulateContentLoaded = () => {
-    // Replace this with actual loading logic
-    // For example, fetch data, load resources, etc.
-    // Once the content is loaded, call setContentLoaded(true)
     setContentLoaded(true);
   };
 
@@ -164,8 +82,7 @@ const CanvasLoader = () => {
               src={loaderImage}
               alt="Loading..."
               style={{
-                // width: "700px", // Customize the image size as needed
-                // height: "400px", // Customize the image size as needed
+               
               }}
             />
             ......................................................
@@ -173,8 +90,6 @@ const CanvasLoader = () => {
         </Html>
       )}
 
-      {/* Replace this button with your actual content */}
-     
  
     </>
   );
