+++ src/components/TabLinks.jsx
@@ -1,32 +1,80 @@
 
-import React from 'react';
+import React from "react";
+import Tilt from "react-tilt";
+import { motion } from "framer-motion";
 import { Link, useNavigate } from 'react-router-dom';
-import { tabData } from "../data.jsx";
-import '../assets/css/TabLinks.css';
 
-const TabComponent = () => {
+import { styles } from "../styles";
+
+import { services } from "../constants";
+import { SectionWrapper } from "../hoc";
+import { fadeIn, textVariant } from "../utils/motion";
+import { meshBounds } from "@react-three/drei";
+
+const ServiceCard = ({ index, title, icon, path }) => {
   const navigate = useNavigate();
 
-  const handleTabClick = (path) => {
+  const handleCardClick = () => {
     navigate(path);
   };
 
   return (
-    <div className='test-container'>
-      <div className="tab-links">
-        {tabData.map((tab, index) => (
-          <div
-            key={index}
-            className="tab-link"
-            style={{ backgroundImage: `url(${tab.backgroundImage})` }}
-            onClick={() => handleTabClick(tab.path)}
-          >
-            {tab.label}
-          </div>
-        ))}
-      </div>
-    </div>
+    <Tilt className='xs:w-[250px] w-full  '>
+      <motion.div
+        variants={fadeIn("right", "spring", index * 0.5, 0.75)}
+        className=' green-pink-gradient p-[1px] rounded-[20px] shadow-card'
+        onClick={handleCardClick} // Call the handleCardClick function on click
+      >
+        <div
+          options={{
+            max: 45,
+            scale: 1,
+            speed: 450,
+          }}
+          className='bg-tertiary rounded-[20px] py-1 px-12 min-h-[40px]  flex justify-evenly items-center flex-col'
+        >
+          <img
+            src={icon}
+            alt={title}
+            className='w-16 h-16 object-contain'
+          />
+
+          <h3 className='text-white text-[20px] font-bold text-center'>
+            {title}
+          </h3>
+        </div>
+      </motion.div>
+    </Tilt>
   );
 };
 
-export default TabComponent;
+const TabLinks = () => {
+  return (
+    <>
+      <motion.div variants={textVariant()}>
+      </motion.div>
+
+      <motion.p
+        variants={fadeIn("", "", 0.1, 1)}
+        className='mt-0   text-secondary text-[50px] max-w-6xl leading-[30px]'
+      >
+ 
+      </motion.p>
+
+      <div className='mt-0 flex  flex-wrap gap-10'>
+        {services.map((services, index) => (
+          <ServiceCard key={services.title} index={index} {...services} />
+        ))}
+      </div>
+    </>
+  );
+};
+
+export default SectionWrapper(TabLinks, "tablinks");
+
+
+
+
+
+
+
