+++ src/components/canvas/EMR_logo.jsx
@@ -9,7 +9,7 @@ import { Canvas } from "@react-three/fiber";
 import { OrbitControls, Preload, meshBounds, useGLTF } from "@react-three/drei";
 
 import CanvasLoader from "../Loader";
-import CanvasLoader2 from "../Loader_Image";
+// import CanvasLoader2 from "../Loader_Image";
 // import { EMR_logo_canvas } from ".";
 
 const EMR_logo = () => {
