+++ src/components/canvas/index.js
@@ -1,5 +1,4 @@
-import EarthCanvas from "./Earth";
 import EMR_logo_canvas from "./EMR_logo";
 import StarsCanvas from "./Stars";
 
-export { EarthCanvas, StarsCanvas,EMR_logo_canvas };
+export { StarsCanvas,EMR_logo_canvas };
