+++ tailwind.config.cjs
@@ -29,16 +29,15 @@ module.exports = {
         // "hero-pattern": "url('/src/assets/hero-bg6.jpg')",
         // "hero-pattern": "url('/src/assets/hero-bg7.jpg')",
         // "hero-pattern": "url('/src/assets/hero-bg8.jpg')",
-        "hero-pattern": "url('/src/assets/hero-bg10.jpg')",
-
-
-
-
-        // "hero-pattern": "url('/src/assets/code-bg.jpg')",
 
         
-
-
+        // "hero-pattern": "url('/src/assets/code-bg.jpg')",
+        
+        
+        
+        
+        // by priyanshu sir 
+        // "hero-pattern": "url('/src/assets/hero-bg10.jpg')",
